Move edit to footer
